requirements.txt
ultralytics
streamlit
pandas
matplotlib
opencv-python
Pillow
numpy
